refactor: extract pre-populate offer logic in its own hook
